New parsing complete. New piping works a charm, issues with parsing args to DOS on stack (int 2fh) all sorted (including in cfg.sys in DOS). Internal commands now use the psp buffer for parsing as this is equivalent to what an external program gets. Code ready to have Environment handling and Batch processing added. DIR now works as specified in the manual. Lots of tape to handle trailing slashes and relative paths... it works and is followable but maybe look at in the future?
COPY needs to be improved to allow for bulk copies, and to allow for all the options.  Perhaps also increase buffer to 512 bytes and dynamically allocated.
REN and DEL need to also be adjusted to work with WC's in the path.

Master ENV needs to be a 160 byte allocated block.