Changed the piping to pull the command line, to allow for AE to actually completely replace the buffer and still proceed. This will be conducive to introducing batch too. Piping being pulled is implemented good!
Implementation of aebuffer is trash and needs work. Decouple the making of pipes from their registration on the cmd line, as a priority to make this code maintainable. Once this decoupling is done, the resultant path of execution being the aebuffer becomes trivial.
Added some stability to the prevent multiple redirouts and redirins from crashing the pipe