Improved the empty command issue handling, removing the ad-hoc solution that handled an empty first command in setupRedir.
Now, before entry into the main execution loop, we check after peeling off redirs if we have any non-<CR> chars on the command line. If not, reset redir vars and reobtain input (either bat file or cmd line). This indicates the command was not accepted.

We now also check for empty commands within a pipeline, that would not be caught by the above (these could arise as the last command in a pipeline). Now instead of such an empty command failing silently, we halt the pipeline, print syntax error and reset.

We also fail if we start a pipe with an empty command (i.e. a command starting with the pipe char).