Command line processing improved. Now trailing slashes and dots and other bad chars are treated properly and robustly.
Moved the setting up of the argument fcbs to the more appropriate place in the command flow.

Changed so that internal commands DONT change the errorlevel (this will be adjusted later).

TODO: Need to add a scanner that sets a flag for "switchchars". If a command doesnt accept switches, we must error with invalid switch error.