Fixed a nasty little bug that could get one stuck in an infinite loop if on a drive where the FAT started failing and the prompt used $p. Now we abort and allow the user to change drives (or change disks before continuing).
Fixed a small bug in bat processing that would not append a CR to the end of the buffer if we reached EOF without a CR in the file. Echo now also doesnt necessarily depend on this CR value. This allows it to deal with "full" lines to echo. Its not always successful.

NEED TO DO: Implement a moratorium on echoing CRLF's after each command in a batch if echo off