Lots of comments and label changes.
Started work on GOTO, which kind of behaves like the main processing loop. Therefore,
opening the batch file has been thrown into a function.

The function which reads a char from the batch file is now is the only place where the EOF state bit is set. This might be moved to the batch main loop if goto doesn't need the flag (though it seems sensible in the grand scheme of things to keep it where it is as it is a legitimate place to set the EOF flag).